Add initial implementation of scarlet-youtube library
- Created Cargo.toml for the scarlet-youtube package.
- Implemented YoutubeSearchResult struct with methods for generating watch URLs.
- Added functions for encoding and parsing search results in TSV format.
- Included utility functions for handling escaped characters in TSV fields.
